pyproject.toml
[tool.poetry]
name = "lectures-gdrive"
version = "0.1.0"
description = ""
authors = ["iamkroot <kroot.patel@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
google-api-python-client = "^1.8.2"
google-auth-oauthlib = "^0.4.1"
requests = "^2.23.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
